money_flow/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-admin-rangefilter==0.13.3
django-smart-selects==1.7.2
python-dotenv==1.1.1
sqlparse==0.5.3
